+++ video-llava/train-1-gpu.sh
@@ -14,5 +14,5 @@ module purge
 singularity exec --nv \
     --overlay /scratch/$USER/my_env/overlay-15GB-500K.ext3:rw \
     /scratch/work/public/singularity/cuda12.3.2-cudnn9.0.0-ubuntu-22.04.4.sif \
-    /bin/bash -c 'source /ext3/env.sh; python train_video_llava.py'
+    /bin/bash -c 'source /ext3/env.sh; python train_model.py'
